Revert "bisiler"
This reverts commit c9086a24240a652fabaae03001eb0c314a50bb82.

+++ lib/services/db.dart
@@ -300,7 +300,7 @@ class DBService {
       description: product["details"],
       sellerToken: product["seller-id"],
       gender: product["gender"] ?? "Unisex",
-      sellerRating: userInfo["rating"].toDouble(),
+      sellerRating: userInfo["rating"],
     );
   }
 
@@ -812,7 +812,8 @@ class DBService {
           username: username,
           rating: reviewInfo["rating"],
           comment: reviewInfo["comment"],
-          reviewDate: reviewInfo["review-date"]);
+          reviewDate: reviewInfo["review-date"]
+      );
     }).toList());
   }
 }
